Survery Paper CSA
Abstract:
Citations have always been an integral part of academic writing and
have given proof of thorough research from time immemorial. Researchers cite
their references to show how meticulously the research has been conducted and
also to avoid plagiarism. Analysis of these citations brings out information
about the quality of papers, research, and institutes. Tracking the frequency of
citation of a paper has been a useful means of its analysis over the years. Although
citation count gives a fair idea about the popularity of the paper, it is persistent
to note that the use of citations by various researchers could also be
made to highlight the imperfection or infirmities. Moreover, it also does not
recognize the assert opinion of the researcher about the citations. This new form
of sentiment detection of citations helps to summarize paper by observing significant
achievements, shortcomings and drawbacks in a particular research
problem. This helps a researcher in understanding the true impact of the paper.
This is where opinion mining meets citation analysis. Algorithms to automatically
detect and infer sentiments from a context around citations are need of the
hour. As a research community, we are poised far behind a human reader for
whom the task of sentiment detection from context is both natural and trivial.
So in this paper, we choose to survey various studies conducted for finding the
context of citation and analyze sentiments there in. Some conclusions from the
survey are also presented at the end.

Keywords:
Citation Analysis, Sentiment Analysis, Citation Context detection.